fix: unexpected exits alert
In the last version of the VEBO contract, the
`getLastRequestedValidatorIndices` method has been removed. Because of
this, the unexpected exits alert stopped working.

Current commit fixes this problem. Now we read all the events of type
`ValidatorExitRequest` emitted by the VEBO contract for the period
specified at the `EXIT_EVENTS_LOOKBACK_WINDOW_IN_SLOTS` parameter in the
`OracleDaemonConfig` contract and store them in the local cache. Then,
information about the validator exits on CL is compared to data in these
events. If the Lido validator is going to exit (a processing slot on CL
has information about it), but VEBO did not request exit of this
validator, the unexpected exit alert is triggered.

Also, remove unnecessary contracts from the application assets and
configs.